+++ torch/nn/functional.py
@@ -1204,9 +1204,10 @@ def instance_norm(input, running_mean, running_var, weight=None, bias=None,
     if bias is not None:
         bias = bias.repeat(b)
 
-    import torch.onnx.symbolic
+    from torch.onnx import symbolic_override_first_arg_based
+    import torch.onnx.core.symbolic
 
-    @torch.onnx.symbolic_override_first_arg_based(torch.onnx.symbolic.instance_norm)
+    @symbolic_override_first_arg_based(torch.onnx.core.symbolic.instance_norm)
     def _instance_norm(input, running_mean=None, running_var=None, weight=None,
                        bias=None, use_input_stats=None, momentum=None, eps=None):
         # Repeat stored stats and affine transform params if necessary
